server/requirements.txt
django==3.1.3
requests
gunicorn==20.1.0
pillow
ibm-cloud-sdk-core==3.10.0
ibm-watson==5.2.2
ibmcloudant==0.0.34
python-dotenv
